removing embed images symbol markdown
The images from copied Jupyter notebook exercises aren't saving in this file, and so a preview of this file doesn't show in GitHub. To try and make this work I will first stop embedding the images and hopefully this will be ignored on GitHub. Otherwise I will erase those lines and hopefully it's fine for the submission.